requirements.txt
pyTelegramBotAPI
yt_dlp
python-dotenv